# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NeutriAI API",
    "description": "AI nutrition, fitness and lifestyle platform. All endpoints except `/healthz`, `/v1/billing/plans` and `/v1/webhooks/stripe` require a Supabase bearer token.",
    "version": "1.0.0"
  },
  "paths": {
    "/v1/me": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "Get Profile",
        "description": "The profile is created by the auth dependency on first authenticated\nrequest, so by the time this runs it exists. ensure_profile is still the\ncall made here rather than a bare select: it keeps this endpoint correct\non its own terms, and it is the one place that returns the deleted-account\n401 if the row is somehow absent.",
        "operationId": "get_profile_v1_me_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "profile"
        ],
        "summary": "Update Profile",
        "operationId": "update_profile_v1_me_patch",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/me/targets": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "Get Targets",
        "operationId": "get_targets_v1_me_targets_get",
        "parameters": [
          {
            "name": "recompute",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Recompute"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TargetsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/me/dashboard": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "Dashboard",
        "description": "One round trip for the entire home screen.",
        "operationId": "dashboard_v1_me_dashboard_get",
        "parameters": [
          {
            "name": "day",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Day"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/me/restrictions": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "List Restrictions",
        "operationId": "list_restrictions_v1_me_restrictions_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "profile"
        ],
        "summary": "Add Restriction",
        "operationId": "add_restriction_v1_me_restrictions_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestrictionIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/me/restrictions/{restriction_id}": {
      "delete": {
        "tags": [
          "profile"
        ],
        "summary": "Remove Restriction",
        "operationId": "remove_restriction_v1_me_restrictions__restriction_id__delete",
        "parameters": [
          {
            "name": "restriction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Restriction Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/me/streaks": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "Get Streaks",
        "operationId": "get_streaks_v1_me_streaks_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/me/history": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "History",
        "operationId": "history_v1_me_history_get",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 365,
              "minimum": 1,
              "default": 30,
              "title": "Days"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/me/push-token": {
      "post": {
        "tags": [
          "profile"
        ],
        "summary": "Register Push Token",
        "description": "Store the device's Expo push token.\n\nCalled on every app launch, not just the first: tokens rotate on reinstall\nand after some OS updates, and a stale token silently stops all delivery.",
        "operationId": "register_push_token_v1_me_push_token_post",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "profile"
        ],
        "summary": "Clear Push Token",
        "operationId": "clear_push_token_v1_me_push_token_delete",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/me/deletion-preview": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "Deletion Preview",
        "description": "What deletion will remove, so the confirmation screen can be specific\nrather than a generic 'are you sure?'.",
        "operationId": "deletion_preview_v1_me_deletion_preview_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/me/account": {
      "delete": {
        "tags": [
          "profile"
        ],
        "summary": "Delete Account",
        "description": "Permanently delete this account and all its data.\n\n``confirm`` must be the literal string DELETE. A destructive, irreversible\naction reached by a single tap is a support ticket waiting to happen.",
        "operationId": "delete_account_v1_me_account_delete",
        "parameters": [
          {
            "name": "confirm",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Confirm"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActivityLevel": {
        "type": "string",
        "enum": [
          "sedentary",
          "light",
          "moderate",
          "active",
          "very_active",
          "athlete"
        ],
        "title": "ActivityLevel"
      },
...
      "Goal": {
        "type": "string",
        "enum": [
          "lose",
          "maintain",
          "gain",
          "recomp"
        ],
        "title": "Goal"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Ok": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "title": "Ok"
      },
...
      "ProfileOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "handle": {
            "type": "string",
            "title": "Handle"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name",
            "default": ""
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "bio": {
            "type": "string",
            "title": "Bio",
            "default": ""
          },
          "sex": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Sex"
              },
              {
                "type": "null"
              }
            ]
          },
          "birth_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Date"
          },
          "height_cm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height Cm"
          },
          "weight_kg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight Kg"
          },
          "target_weight_kg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Weight Kg"
          },
          "activity_level": {
            "$ref": "#/components/schemas/ActivityLevel",
            "default": "moderate"
          },
          "goal": {
            "$ref": "#/components/schemas/Goal",
            "default": "maintain"
          },
          "diet_mode": {
            "$ref": "#/components/schemas/DietMode",
            "default": "balanced"
          },
          "unit_system": {
            "type": "string",
            "title": "Unit System",
            "default": "imperial"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone",
            "default": "UTC"
          },
          "is_private": {
            "type": "boolean",
            "title": "Is Private",
            "default": false
          },
          "onboarded_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Onboarded At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "handle"
        ],
        "title": "ProfileOut"
      },
...
      "RestrictionIn": {
        "properties": {
          "kind": {
            "type": "string",
            "pattern": "^(allergy|intolerance|avoid|religious|preference)$",
            "title": "Kind",
            "default": "allergy"
          },
          "label": {
            "type": "string",
            "maxLength": 60,
            "minLength": 1,
            "title": "Label"
          },
          "severity": {
            "type": "string",
            "pattern": "^(mild|moderate|severe|anaphylactic)$",
            "title": "Severity",
            "default": "moderate"
          }
        },
        "type": "object",
        "required": [
          "label"
        ],
        "title": "RestrictionIn"
      },
...
      "Sex": {
        "type": "string",
        "enum": [
          "male",
          "female",
          "other"
        ],
        "title": "Sex"
      },
...
      "TargetsOut": {
        "properties": {
          "bmr_kcal": {
            "type": "integer",
            "title": "Bmr Kcal"
          },
          "tdee_kcal": {
            "type": "integer",
            "title": "Tdee Kcal"
          },
          "target_kcal": {
            "type": "integer",
            "title": "Target Kcal"
          },
          "protein_g": {
            "type": "integer",
            "title": "Protein G"
          },
          "carbs_g": {
            "type": "integer",
            "title": "Carbs G"
          },
          "fat_g": {
            "type": "integer",
            "title": "Fat G"
          },
          "fiber_g": {
            "type": "integer",
            "title": "Fiber G"
          },
          "sugar_g_max": {
            "type": "integer",
            "title": "Sugar G Max"
          },
          "water_ml": {
            "type": "integer",
            "title": "Water Ml"
          },
          "rationale": {
            "additionalProperties": true,
            "type": "object",
            "title": "Rationale",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "bmr_kcal",
          "tdee_kcal",
          "target_kcal",
          "protein_g",
          "carbs_g",
          "fat_g",
          "fiber_g",
          "sugar_g_max",
          "water_ml"
        ],
        "title": "TargetsOut"
      },
...
    }
  }
}